Add Comparator to reduce duplicate code
